added stamina drain and recharge and dodge

+++ Assets/Scripts/Managers/CharacterStats.cs
@@ -13,8 +13,8 @@ namespace CW
         public int currentHealth;
 
         public int staminaLevel = 10;
-        public int maxStamina;
-        public int currentStamina;
+        public float maxStamina;
+        public float currentStamina;
 
         public bool isDead;
